requirements.txt
Flask
kafka-python
pyspark
